bcc_detection/requirements.txt
# Deep Learning
torch>=2.0.0
torchvision>=0.15.0

# Data Processing
numpy>=1.21.0
pandas>=1.3.0
scikit-learn>=1.0.0
scipy>=1.7.0
scikit-image>=0.19.0
opencv-python>=4.5.0
Pillow>=9.0.0
tqdm>=4.62.0

# Visualization
matplotlib>=3.5.0
seaborn>=0.11.0

# Utilities
pyyaml>=5.4.1
tifffile>=2021.7.2
openslide-python>=1.1.2
albumentations>=1.3.0
tensorboard>=2.8.0
wandb>=0.12.0
pytest>=7.0.0
black>=22.0.0
flake8>=4.0.0
isort>=5.10.0
mypy>=0.910 